changed some auth things

+++ server/primalformulas/views.py
@@ -1,7 +1,6 @@
-from typing import cast
-
 from rest_framework import status
-from django.contrib.auth.models import AbstractBaseUser, AnonymousUser, User
+
+from django.contrib.auth.models import AbstractBaseUser, AnonymousUser
 from rest_framework.permissions import AllowAny, IsAuthenticated
 from rest_framework.request import Request
 from rest_framework.response import Response
@@ -9,7 +8,8 @@ from rest_framework.views import APIView
 from rest_framework.authtoken.models import Token
 from rest_framework.authentication import SessionAuthentication, TokenAuthentication
 
-from django.contrib.auth import authenticate
+from django.db import transaction
+from django.contrib.auth import authenticate, login, logout
 
 from primalformulas.serializers import UserSerializer
 
@@ -41,14 +41,16 @@ class RegisterView(APIView):
 
     def post(self, request: Request) -> Response:
         serializer = UserSerializer(data=request.data)
-
         if not serializer.is_valid():
             return Response(serializer.errors, status=status.HTTP_401_UNAUTHORIZED)
 
-        user = cast(User, serializer.save())
-        user.set_password(request.data["password"])
+        with transaction.atomic():
+            user = serializer.save()
+            user.set_password(request.data["password"])
+            user.save()
 
-        token, _ = Token.objects.get_or_create(user=user)
+            token, _ = Token.objects.get_or_create(user=user)
+
         return Response(
             {"token": token.key, "user": serializer.data},
             status=status.HTTP_201_CREATED,
@@ -60,21 +62,22 @@ class LoginView(APIView):
     authentication_classes = [TokenAuthentication, SessionAuthentication]
 
     def post(self, request: Request) -> Response:
-        try:
-            username = request.data["username"]
-            password = request.data["password"]
-        except KeyError:
+        username = request.data["username"]
+        password = request.data["password"]
+
+        if not username or not password:
             return Response(
                 {"error": "Username and password are required."},
                 status=status.HTTP_400_BAD_REQUEST,
             )
 
         user = authenticate(username=username, password=password)
-
         if user is None:
             return Response(
                 {"error": "Invalid Credentials"}, status=status.HTTP_401_UNAUTHORIZED
             )
+        else:
+            login(request, user)
 
         token, _ = Token.objects.get_or_create(user=user)
         serializer = UserSerializer(instance=user)
@@ -92,7 +95,11 @@ class LogoutView(APIView):
                 status=status.HTTP_400_BAD_REQUEST,
             )
 
-        request.user.auth_token.delete()
+        if hasattr(request.user, "auth_token"):
+            request.user.auth_token.delete()
+
+        if request.user:
+            logout(request)
 
         return Response(
             {"Message": "User successfully logged out"},
